que inicie el intercambio despues de deffense
